% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AntibodyForests_data.R
\docType{data}
\name{af_default}
\alias{af_default}
\title{Small AntibodyForests object with default algorithm for function testing purposes}
\format{
An object of class \code{list} of length 1.
}
\usage{
af_default
}
\description{
Small AntibodyForests object with default algorithm for function testing purposes
}
\keyword{datasets}
