% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AntibodyForests_data.R
\docType{data}
\name{compare_repertoire}
\alias{compare_repertoire}
\title{Example output from Af_compare_within_repertoires() for function testing purposes}
\format{
An object of class \code{list} of length 1.
}
\usage{
compare_repertoire
}
\description{
Example output from Af_compare_within_repertoires() for function testing purposes
}
\keyword{datasets}
