% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraph_to_phylo.R
\name{igraph_to_phylo}
\alias{igraph_to_phylo}
\title{Converts an igraph network into a phylogenetic tree as a phylo object.}
\usage{
igraph_to_phylo(tree, solve_multichotomies)
}
\arguments{
\item{tree}{igraph object}

\item{solve_multichotomies}{boolean - whether to remove multichotomies in the resulting phylogenetic tree using ape::multi2di}
}
\value{
phylogenetic tree
}
\description{
Converts an igraph network into a phylogenetic tree as a phylo object.
}
\examples{
phylo_object <- igraph_to_phylo(AntibodyForests::small_af[["S1"]][["clonotype1"]]$igraph)
}
