% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AntibodyForests_data.R
\docType{data}
\name{PLM_dataframe}
\alias{PLM_dataframe}
\title{Small PLM dataframe for function testing purposes}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 60 rows and 11 columns.
}
\usage{
PLM_dataframe
}
\description{
Small PLM dataframe for function testing purposes
}
\keyword{datasets}
