% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AntibodyForests_data.R
\docType{data}
\name{small_af}
\alias{small_af}
\title{Small AntibodyForests object for function testing purposes}
\format{
An object of class \code{AntibodyForests} of length 5.
}
\usage{
small_af
}
\description{
Small AntibodyForests object for function testing purposes
}
\keyword{datasets}
