% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AntibodyForests_data.R
\docType{data}
\name{small_vdj}
\alias{small_vdj}
\title{Small VDJ dataframe for function testing purposes}
\format{
An object of class \code{data.frame} with 3671 rows and 70 columns.
}
\usage{
small_vdj
}
\description{
Small VDJ dataframe for function testing purposes
}
\keyword{datasets}
